ANDROID: KVM: arm64: Read REVIDR_EL1 and AIDR_EL1 for protected VMs
Access to REVIDR_EL1 and AIDR_EL1 is implementation defined. Revert the
behavior of reading REVIDR_EL1 to RAZ/WI, and do the same for AIDR_EL1,
instead of being handled by the host.

In older kernels, handling by the host of accesses to REVIDR_EL1 and
AIDR_EL1 is broken, since the sysreg configuration information wasn't
being populated for them. This has been fixed upstream in commit
4cd48565b0e5 ("KVM: arm64: Set HCR_EL2.TID1 unconditionally"). However,
that fix is rather invasive, and could potentially cause a change to the
values existing protected guests would see for these register, since up
until recently, these registers have either been handled as RAZ, or not
exposed to protected guests to begin with.

Fixes: 81d453debbb2 ("BACKPORT: FROMGIT: KVM: arm64: Handle AIDR_EL1 and REVIDR_EL1 in host for protected VMs")
Bug: 446864228
Bug: 446746493
Change-Id: Iec63c748b0561ab58563b92cc8ea4c2950c31aef
Signed-off-by: Fuad Tabba <tabba@google.com>