ANDROID: vendor_hooks: Modify the location of trace_android_vh_nr_pcp_alloc
Since trace_android_vh_rmqueue_bulk_bypass uses the reserved pool to fill when pcp is empty, in this case, the dynamic modification of the batch by trace_android_vh_nr_pcp_alloc is meaningless. Therefore, we modify the position of trace_android_vh_nr_pcp_alloc and dynamically change the batch when the reserved pool fails to be filled.

Bug: 438620264

Change-Id: I1c5c397bd0874a2e17554eea2feca4c678ef9162
Signed-off-by: Marcus Ma <maminghui5@xiaomi.corp-partner.google.com>